Corrected issue with missing References after 4.5 upgrade
Apparently in .net 4.5 there are multiple libraries that have been
pushed off now to Nuget so they can be worked on and released faster
than the .net 4.5 libaraies.
System.Web.Optimization was the main culprit here.  Also the OpenAuth
libraries were not there, but i was able to add back through nuget.